Refactor image opacity and background styling
This commit refactors the code related to image opacity and background styling in the sports.html file. The changes include updating the opacity of images and adding or removing the "blank" class from the wrapper element to control the background display. These changes improve the visibility and appearance of the images on the webpage.